feat(api): move end point to etablissemnt services
