+++ packages/api/src/modules/associations/associations.service.test.ts
@@ -294,76 +294,4 @@ describe("AssociationService", () => {
             expect(actual).toEqual(expected)
         })
     })
-
-    describe("getEtablissement()", () => {
-        const NIC = "00032"
-        it("should call etablissementService.getEtablissement()", async () => {
-            getEtablissementMock.mockImplementationOnce(() => Promise.resolve({ etablissement: true } as unknown as Etablissement));
-            getIdentifierTypeMock.mockImplementationOnce(() => StructureIdentifiersEnum.siren);
-            await associationsService.getEtablissement(SIREN, NIC);
-            expect(getEtablissementMock).toHaveBeenCalledWith(SIREN + NIC);
-        })
-
-        it("should call search siren match with rna", async () => {
-            const expected = SIREN;
-            getEtablissementMock.mockImplementationOnce(() => Promise.resolve({ etablissement: true } as unknown as Etablissement));
-            getIdentifierTypeMock.mockImplementationOnce(() => StructureIdentifiersEnum.rna);
-            rnaSirenServiceGetSirenMock.mockImplementationOnce(() => Promise.resolve(expected));
-            await associationsService.getEtablissement(SIREN, NIC);
-            expect(getEtablissementMock).toHaveBeenCalledWith(expected + NIC);
-        })
-
-
-        it("should throw error not found error (siren not matching with rna)", async () => {
-            const expected = `We haven't found a corresponding SIREN to the given RNA ${SIREN}`
-
-            rnaSirenServiceGetSirenMock.mockImplementationOnce(() => Promise.resolve(null));
-            getIdentifierTypeMock.mockImplementationOnce(() => StructureIdentifiersEnum.rna);
-            let actual;
-            try {
-                actual = await associationsService.getEtablissement(SIREN, NIC);
-            } catch (e) {
-                actual = (e as Error).message
-            }
-            expect(actual).toEqual(expected);
-        })
-
-        it("should throw error not found error (EtablissementService return null)", async () => {
-            const expected = "Etablissement not found"
-
-            getEtablissementMock.mockImplementationOnce(() => Promise.resolve(null));
-            getIdentifierTypeMock.mockImplementationOnce(() => StructureIdentifiersEnum.siren);
-            let actual;
-            try {
-                actual = await associationsService.getEtablissement(SIREN, NIC);
-            } catch (e) {
-                actual = (e as Error).message
-            }
-            expect(actual).toEqual(expected);
-        })
-
-        it("should throw error (identifiers type not accepted)", async () => {
-            const expected = "You must provide a valid SIREN or RNA";
-            getIdentifierTypeMock.mockImplementationOnce(() => StructureIdentifiersEnum.siret);
-            let actual;
-            try {
-                actual = await associationsService.getEtablissement(SIREN, NIC);
-            } catch (e) {
-                actual = (e as Error).message
-            }
-            expect(actual).toEqual(expected);
-        })
-
-        it("should throw error (identifiers type not found)", async () => {
-            const expected = "You must provide a valid SIREN or RNA";
-            getIdentifierTypeMock.mockImplementationOnce(() => null);
-            let actual;
-            try {
-                actual = await associationsService.getEtablissement(SIREN, NIC);
-            } catch (e) {
-                actual = (e as Error).message
-            }
-            expect(actual).toEqual(expected);
-        })
-    })
 });
